+++ weaver/datatype.py
@@ -822,7 +822,23 @@ class Process(Base):
     @property
     def inputs(self):
         # type: () -> Optional[List[Dict[str, Any]]]
-        return self.get("inputs")
+        """
+        According to `OGC-API`, ``maxOccurs`` and ``minOccurs`` representations should be:
+            - ``maxOccurs``: ``int`` or ``"unbounded"``
+            - ``minOccurs``: ``int``
+
+        .. note::
+            Because of pre-existing/deployed/remote processes, inputs are formated to respect the valid representation.
+        """
+
+        inputs = self.get("inputs")
+        if inputs is not None:
+            for input_ in inputs:
+                input_["minOccurs"] = int(input_["minOccurs"])
+                input_["maxOccurs"] = (
+                    int(input_["maxOccurs"]) if input_["maxOccurs"] != "unbounded" else input_["maxOccurs"]
+                )
+        return inputs
 
     @property
     def outputs(self):